+++ CHANGES.md
@@ -1,6 +1,11 @@
 # Changes
 
-## 2.0.0 (dev)
+## 2.0.0-alpha.2 (dev)
+
+* Replace pysam with bamnostic
+* Re-enable testing on Python 3.8
+
+## 2.0.0-alpha.1 (2019-12-25)
 
 * *Breaking changes:*
   * Dropped support for python 3.3-3.5 in order to take advantage of many new features in 3.6 (such as type annotations), and to migrate to xphyle for file management. The 1.1.x branch will maintain 3.3-3.5 support and will receive any new bug fixes (but no new features).